requirements.txt
# Core FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Environment variables
python-dotenv>=1.0.0

# AWS SDK
boto3>=1.26.0
botocore>=1.29.0

# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.5.0

# Core scientific computing
numpy>=1.21.0
requests>=2.25.0

# Transformers and HuggingFace
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0
huggingface-hub>=0.16.0

# Configuration management (for EVF-SAM)
hydra-core>=1.3.0

# Additional ML dependencies
einops>=0.6.0

# Florence-2 specific dependencies
timm>=0.9.0
sentencepiece>=0.1.99

# Optional performance optimizations (may fail on some systems)
# flash-attn>=2.0.0
# xformers>=0.0.20
